% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-assert.R
\name{assert_df}
\alias{assert_df}
\title{Check that df has x column names}
\usage{
assert_df(df, required_columns, .env = rlang::caller_env())
}
\arguments{
\item{df}{df to check}

\item{required_columns}{required column names}

\item{.env}{caller environment to complain from}
}
\value{
silent if ok or else an error if something is missing.
}
\description{
Check that df has x column names
}
\keyword{internal}
