% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09-autoplot.R
\name{autoplot.ff_simulation}
\alias{autoplot.ff_simulation}
\alias{plot.ff_simulation}
\title{Automatically Plot ff_simulation Object}
\usage{
autoplot.ff_simulation(object, type = c("wins", "rank", "points"), ...)

\method{plot}{ff_simulation}(x, ..., type = c("wins", "rank", "points"), y)
}
\arguments{
\item{object}{a \code{ff_simulation} object as created by \code{ff_simulate()}}

\item{type}{one of "wins", "rank", "points"}

\item{...}{unused, required by autoplot generic}

\item{x}{A \code{ff_simulation} object.}

\item{y}{Ignored, required for compatibility with the \code{plot()} generic.}
}
\value{
a ggplot object
}
\description{
Creates automatic plots for wins, ranks, or points for an \code{ff_simulation} object as created by \code{ff_simulate()}.
}
\details{
Usage of this function/method requires the ggplot2 package and (for wins and points plots) the ggridges package.
}
\examples{
\donttest{

simulation <- .ffs_cache("foureight_sim.rds")

ggplot2::autoplot(simulation) # default is type = "wins"
ggplot2::autoplot(simulation, type = "rank")
ggplot2::autoplot(simulation, type = "points")
}

}
\seealso{
\code{vignette("basic")} for example usage
}
