% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data.R
\name{.ffs_read_data}
\alias{.ffs_read_data}
\alias{.ffs_cache_dir}
\title{Read data from cache or pkg data}
\usage{
.ffs_read_data(file_name, .env = rlang::caller_env())

.ffs_cache_dir(...)
}
\arguments{
\item{file_name}{filename to find}

\item{.env}{for internal use in error messages}
}
\description{
Read data from cache or pkg data
}
