% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_outcomes.R
\name{.ff_adp_outcomes}
\alias{.ff_adp_outcomes}
\title{Connects ff_scoringhistory to past rankings}
\usage{
.ff_adp_outcomes(scoring_history, injury_model)
}
\value{
dataframe with position, rank, probability of games played, all week score outcomes
}
\description{
Connects ff_scoringhistory to past rankings
}
