% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_outcomes.R
\name{.ff_apply_gp_model}
\alias{.ff_apply_gp_model}
\title{Applies various injury models to adp outcomes}
\usage{
.ff_apply_gp_model(adp_outcomes, model_type)
}
\value{
same adp outcomes dataframe but with a prob_gp column
}
\description{
Applies various injury models to adp outcomes
}
\keyword{internal}
