% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ff_simulate.R
\name{.ff_apply_injury_model}
\alias{.ff_apply_injury_model}
\title{Applies various injury models to adp outcomes}
\usage{
.ff_apply_injury_model(adp_outcomes, model_type)
}
\description{
Applies various injury models to adp outcomes
}
