% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schedule.R
\name{.ff_build_schedules}
\alias{.ff_build_schedules}
\title{Generate fantasy schedules}
\usage{
.ff_build_schedules(n_teams = 14, n_seasons = 100, n_weeks = 13, seed = NULL)
}
\arguments{
\item{n_teams}{number of teams in simulation, default = 14}

\item{n_seasons}{number of seasons to simulate, default = 100}

\item{seed}{an integer to control reproducibility}

\item{weeks_per_season}{number of weeks per season, default = 17}
}
\description{
The main function of the package
}
