% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_preprocessing.R
\name{.ff_join_data.sleeper_conn}
\alias{.ff_join_data.sleeper_conn}
\title{sleeper joins}
\usage{
\method{.ff_join_data}{sleeper_conn}(conn, rosters, latest_rankings, adp_outcomes)
}
\description{
sleeper joins
}
