% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6_optimise.R
\name{.ff_optimise_one_lineup}
\alias{.ff_optimise_one_lineup}
\title{Optimise single lineup}
\usage{
.ff_optimise_one_lineup(franchise_scores, lineup_constraints)
}
\arguments{
\item{franchise_scores}{a data frame of scores for one week and one franchise}

\item{lineup_constraints}{a data frame as created by \code{ffscrapr::ff_starter_positions()}}
}
\value{
a list including the optimal_score and the optimal_lineup.
}
\description{
Optimises lineups for one franchise week at a time. Use purrr or loops to do more franchises/weeks/seasons
}
\keyword{internal}
