% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_optimize.R
\name{.ff_optimize_lineups}
\alias{.ff_optimize_lineups}
\title{Optimize Lineups}
\usage{
.ff_optimize_lineups(
  projected_scores,
  lineup_constraints,
  best_ball = TRUE,
  parallel = FALSE
)
}
\description{
Optimizes lineups for all franchises in the dataframe
}
