% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/15-deal-scores.R
\name{.ffs_grade}
\alias{.ffs_grade}
\title{Letter grade for a deal score}
\usage{
.ffs_grade(s)
}
\arguments{
\item{s}{a score in equivalent playoff percentage points}
}
\description{
Letter grade for a deal score
}
\keyword{internal}
