% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-ffscrapr.R
\name{ff_rosters}
\alias{ff_rosters}
\title{Get league rosters}
\description{
See \code{ffscrapr::ff_rosters} for details.
}
\seealso{
Other ffscrapr-imports: 
\code{\link{ff_connect}()},
\code{\link{ff_scoringhistory}()},
\code{\link{ff_starter_positions}()}
}
\concept{ffscrapr-imports}
