% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-ffscrapr.R
\name{ff_starter_positions}
\alias{ff_starter_positions}
\title{Get league starter positions}
\description{
See \code{ffscrapr::ff_starter_positions} for details.
}
\seealso{
Other ffscrapr-imports: 
\code{\link{ff_connect}()},
\code{\link{ff_scoringhistory}()}
}
\concept{ffscrapr-imports}
