% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-rosters_replacementlevel.R
\name{ffs_add_replacement_level}
\alias{ffs_add_replacement_level}
\title{Add replacement level players to each roster}
\usage{
ffs_add_replacement_level(
  rosters,
  latest_rankings,
  franchises,
  lineup_constraints,
  pos_filter = c("QB", "RB", "WR", "TE")
)
}
\arguments{
\item{rosters}{a dataframe of rosters as created by \code{ffs_rosters()}}

\item{latest_rankings}{a dataframe of latest rankings as created by \code{ff_latest_rankings()}}

\item{franchises}{a dataframe of franchises as created by \code{ffs_franchises()}}

\item{lineup_constraints}{a dataframe of lineup constraints as created by \code{ffs_starter_positions}}

\item{pos_filter}{a character vector of positions to filter to, defaults to c("QB","RB","WR","TE","K")}
}
\value{
a dataframe of rosters with replacements
}
\description{
Add replacement level players to each roster
}
