% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6_schedule.R
\name{ffs_build_schedules}
\alias{ffs_build_schedules}
\title{Generate fantasy schedules}
\usage{
ffs_build_schedules(n_teams = 14, n_seasons = 100, n_weeks = 13, seed = NULL)
}
\arguments{
\item{n_teams}{number of teams in simulation, default = 14}

\item{n_seasons}{number of seasons to simulate, default = 100}

\item{n_weeks}{number of weeks per season, default = 17}

\item{seed}{an integer to control reproducibility}
}
\description{
The main function of the package
}
