% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_copy.R
\name{ffs_copy_template}
\alias{ffs_copy_template}
\title{Copy simulation template to filename}
\usage{
ffs_copy_template(
  filename = "ff_simulation.R",
  template = c("season", "rest_of_season", "week"),
  overwrite = NULL
)
}
\arguments{
\item{filename}{New file name, defaults to putting "ff_simulation.R" into your current directory}

\item{template}{choice of template: one of "season", "week"}

\item{overwrite}{a logical (or NULL) - overwrite if existing file found?}
}
\value{
a success message signalling success/failure.
}
\description{
Creates a simulation template file with all of the components of ff_simulate, ready for tinkering!
}
\examples{
\donttest{

tmp <- tempfile()
ffs_copy_template(tmp)
}
}
