% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-rosters.R
\name{ffs_franchises}
\alias{ffs_franchises}
\title{Get Franchises}
\usage{
ffs_franchises(conn)
}
\arguments{
\item{conn}{a connection object as created by \code{ffscrapr::ff_connect()} and friends.}
}
\value{
a dataframe of franchises that includes the league_id column
}
\description{
This function lightly wraps \code{ffscrapr::ff_franchises()} and adds league_id, which is a required column for ffsimulator.
}
\examples{
\donttest{
# cached examples
conn <- .ffs_cache("mfl_conn.rds")

try({ # prevents CRAN connectivity issues, not actually required in normal usage
ffs_franchises(conn)
})
}

}
\seealso{
vignette("Custom Simulations") for more detailed example usage
}
