% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_projections.R
\name{ffs_generate_predictions}
\alias{ffs_generate_predictions}
\title{Generate Predictions}
\usage{
ffs_generate_predictions(preprocessed_data, n_seasons, n_weeks)
}
\arguments{
\item{preprocessed_data}{a dataframe as created by \code{ffs_preprocess_data}}

\item{n_seasons}{number of seasons}

\item{n_weeks}{weeks per season}
}
\value{

}
\description{
Run the bootstrapped resampling of player week outcomes on preprocessed dataframes for a given number of seasons and weeks per season.
}
\details{
Dataframe should contain ecr, prob_gp, week_outcomes, injury_model, franchise_id, pos, and bye
}
