% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_rankings.R
\name{ffs_latest_rankings}
\alias{ffs_latest_rankings}
\title{Download latest rankings from DynastyProcess GitHub}
\usage{
ffs_latest_rankings()
}
\description{
Fetches a copy of FP data from DynastyProcess's data repository.
}
\details{
If you have any issues with the output of this data, please open an issue in
the DynastyProcess data repository.
}
\seealso{
\url{https://github.com/dynastyprocess/data}
}
