% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_optimize.R
\name{ffs_optimize_lineups}
\alias{ffs_optimize_lineups}
\alias{ffs_optimise_lineups}
\title{Optimize Lineups}
\usage{
ffs_optimize_lineups(
  projected_scores,
  lineup_constraints,
  best_ball = TRUE,
  parallel = FALSE
)

ffs_optimise_lineups(
  projected_scores,
  lineup_constraints,
  best_ball = TRUE,
  parallel = FALSE
)
}
\description{
Optimizes lineups for all franchises in the dataframe
}
