% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07-schedule.R
\name{ffs_repeat_schedules}
\alias{ffs_repeat_schedules}
\title{Repeat fantasy schedules}
\usage{
ffs_repeat_schedules(actual_schedule, n_seasons)
}
\arguments{
\item{actual_schedule}{a schedule retrieved by \code{ffs_schedule()}}

\item{n_seasons}{number of seasons to simulate, default = 100}
}
\value{
a dataframe of schedules for the simulation
}
\description{
This function repeats an actual \code{ffs_schedule()} by the appropriate number of seasons.
}
\examples{
\donttest{
try({ # try block to prevent CRAN-related issues
conn <- .ffs_cache_example("mfl_conn.rds") # cached connection
actual_schedule <- ffs_schedule(conn)

ffs_repeat_schedules(actual_schedule = actual_schedule, n_seasons = 10)
})
}

}
\seealso{
\code{vignette("Custom Simulations")} for example usage
}
