% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07-schedule.R
\name{ffs_schedule}
\alias{ffs_schedule}
\title{Get Schedule}
\usage{
ffs_schedule(conn)
}
\arguments{
\item{conn}{a connection object as created by \code{ffscrapr::ff_connect()} and friends.}
}
\value{
a dataframe of schedule that includes the league_id column
}
\description{
This function lightly wraps \code{ffscrapr::ff_schedule()} and adds league_id, which is a required column for ffsimulator, casts IDs to character, and drops actual games played so as to only simulate unplayed games.
}
\examples{
\donttest{
# cached examples
try({ # try block to prevent CRAN-related issues
conn <- .ffs_cache("mfl_conn.rds")
ffs_schedule(conn)
})
}

}
\seealso{
vignette("Custom Simulations") for more detailed example usage
}
