% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-ffscrapr.R
\name{ffs_starter_positions}
\alias{ffs_starter_positions}
\title{Get league starter positions}
\usage{
ffs_starter_positions(conn)
}
\arguments{
\item{conn}{a connection object as created by \code{ffscrapr::ff_connect()} and friends.}
}
\value{
A tidy dataframe of positional lineup rules, one row per position with minimum and maximum starters as well as total starter calculations.
}
\description{
This function lightly wraps \code{ffscrapr::ff_starter_positions()} and cleans up some abbreviations (PK -> K)
}
\examples{
\donttest{
# cached examples
conn <- .ffs_cache("mfl_conn.rds")

ffs_starter_positions(conn)
}

}
