% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data.R
\name{ffs_update_data_template}
\alias{ffs_update_data_template}
\title{Copy template for updating rankings}
\usage{
ffs_update_data_template(filename = "ffsim_update_data.R", overwrite = NULL)
}
\arguments{
\item{filename}{New file name, defaults to putting "ff_simulation.R" into your current directory}

\item{overwrite}{a logical (or NULL) - overwrite if existing file found?}
}
\value{
a success message signalling success/failure.
}
\description{
Creates a template file for updating package rankings. This involves some
devel packages eg ffpros, so is only provided as a script rather than as a
package function.
}
\examples{
\donttest{
tmp <- tempfile()
ffs_update_data_template(tmp)
}
}
