% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{package}
\name{ffsimulator-package}
\alias{ffsimulator}
\alias{ffsimulator-package}
\title{ffsimulator: Simulate Fantasy Football Seasons}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Simulate fantasy football seasons.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tanho63/ffsimulator}
  \item Report bugs at \url{https://github.com/tanho63/ffsimulator/issues}
}

}
\author{
\strong{Maintainer}: Tan Ho \email{tan@tanho.ca} (\href{https://orcid.org/0000-0001-8388-5155}{ORCID})

}
\keyword{internal}
