% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-ffscrapr.R
\name{fleaflicker_connect}
\alias{fleaflicker_connect}
\title{Connect to a league}
\description{
See \code{ffscrapr::fleaflicker_connect()} for details.
}
\seealso{
Other ffscrapr-imports: 
\code{\link{espn_connect}()},
\code{\link{ff_connect}()},
\code{\link{ff_scoringhistory}()},
\code{\link{ff_starter_positions}()},
\code{\link{mfl_connect}()},
\code{\link{sleeper_connect}()}
}
\concept{ffscrapr-imports}
