% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data.R
\name{fp_dynasty_history}
\alias{fp_dynasty_history}
\title{Historical dynasty ranks}
\usage{
fp_dynasty_history()
}
\value{
dataframe of preseason dynasty rankings from ffpros
}
\description{
This dataframe has historical FantasyPros dynasty-overall rankings starting
in 2018 for QB/RB/WR/TE, including player age at the time of ranking, as
gathered by the ffpros package. Used to model year-over-year dynasty value
transitions.
}
\details{
By default, it returns bundled package data unless the \code{ffsimulator.cache_directory}
option is configured, in which case it will look for a rds file named
\code{fp_dynasty_history.rds} in that directory.

To update, use \code{ffs_update_data_template()} to find the bundled update script
}
