% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data.R
\name{fp_injury_table}
\alias{fp_injury_table}
\title{FP injury table}
\usage{
fp_injury_table()
}
\value{
dataframe of injury probability by position rank
}
\description{
This dataframe contains a column (\code{prob_gp}) for each positional ranking that
describes the probability of a player with that preseason ADP playing in a
given game. It is modelled from historical rankings data and the number of
games played per season for a given positional rank.
}
\details{
By default, it returns bundled package data unless the \code{ffsimulator.cache_directory}
option is configured, in which case it will look for a rds file named
\code{fp_injury_table.rds} in that directory.
}
