% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_outcomes.R
\docType{data}
\name{fp_rankings_history}
\alias{fp_rankings_history}
\title{Historical position ranks}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 7212 rows and 10 columns.
}
\usage{
fp_rankings_history
}
\description{
This dataframe has historical positional rankings for 2012:2020 QB/RB/WR/TE and 2015-2020 DL/LB/DB.
}
\keyword{datasets}
