% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data.R
\name{fp_rankings_history_week}
\alias{fp_rankings_history_week}
\title{Historical position ranks}
\usage{
fp_rankings_history_week()
}
\value{
dataframe of in-season positional rankings
}
\description{
This dataframe has historical positional in-season rankings starting in 2012 for
QB/RB/WR/TE/PK and 2015 for DL/LB/DB, as gathered by the ffpros package.
}
\details{
By default, it returns bundled package data unless the \code{ffsimulator.cache_directory}
option is configured, in which case it will look for a rds file named
\code{fp_rankings_history_week.rds} in that directory.
}
