% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-ffscrapr.R
\name{sleeper_connect}
\alias{sleeper_connect}
\title{Connect to a league}
\value{
a connection object to be used with \verb{ff_*} functions
}
\description{
See \code{ffscrapr::sleeper_connect()} for details.
}
\seealso{
Other ffscrapr-imports: 
\code{\link{espn_connect}()},
\code{\link{ff_connect}()},
\code{\link{ff_scoringhistory}()},
\code{\link{ff_starter_positions}()},
\code{\link{fleaflicker_connect}()},
\code{\link{mfl_connect}()}
}
\concept{ffscrapr-imports}
