% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_preprocessing.R
\name{.ff_join_data}
\alias{.ff_join_data}
\title{Join all the data together for preprocessing}
\usage{
.ff_join_data(conn, rosters, latest_rankings, adp_outcomes)
}
\description{
Join all the data together for preprocessing
}
\keyword{internal}
