% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ff_simulate.R
\name{.ff_optimize_lineups}
\alias{.ff_optimize_lineups}
\title{Optimize Lineups}
\usage{
.ff_optimize_lineups(franchise_scores, lineup_constraints)
}
\arguments{
\item{franchise_scores}{a data frame of scores for one week and one franchise}

\item{lineup_constraints}{a data frame as created by \code{ffscrapr::ff_starter_positions()}}
}
\value{
a list including the optimal_score and the optimal_lineup tibble.
}
\description{
Optimizes lineups for one franchise week at a time. Use purrr or loop to do more franchises/weeks/seasons
}
