% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-outcomes.R
\name{.ff_rank_expand}
\alias{.ff_rank_expand}
\title{Expand one rank into a vector of five ranks to broaden population of possible outcomes}
\usage{
.ff_rank_expand(.x)
}
\description{
Expand one rank into a vector of five ranks to broaden population of possible outcomes
}
\keyword{internal}
