% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ff_simulate.R
\name{ff_simulate}
\alias{ff_simulate}
\title{Simulate Fantasy Seasons}
\usage{
ff_simulate(
  conn,
  n_seasons = 100,
  weeks_per_season = 17,
  best_ball = FALSE,
  seed = NULL,
  custom_rankings = NULL,
  injury_model = c("simple", "none"),
  owner_efficiency = NULL,
  base_seasons = 2016:2020
)
}
\arguments{
\item{conn}{an connection to a league made with \code{ff_connect()} and friends (required)}

\item{n_seasons}{number of seasons to simulate, default = 100}

\item{weeks_per_season}{number of weeks per season, default = 17}

\item{best_ball}{Are weekly wins based on optimal score?}

\item{seed}{an integer to control reproducibility}

\item{custom_rankings}{a dataframe with x specification to pass in as latest fantasy rankings}

\item{injury_model}{select between "bimodal", "none" - later we'll have separate}

\item{owner_efficiency}{pass in a named list with average and sd}

\item{verbose}{print progress messages for debugging}
}
\description{
The main function of the package
}
\examples{
\dontrun{

conn <- mfl_connect(2021, 54040)

auto <- ff_simulate(conn)

reprex <- ff_simulate(conn = conn, seed = 613)

basic <- ff_simulate(conn = conn, n_seasons = 100, weeks_per_season = 17, best_ball = FALSE)

custom <- ff_simulate(
  conn = conn,
  n_seasons = 100,
  weeks_per_season = 17,
  custom_rankings = df_rankings,
  seed = 613,
  best_ball = FALSE,
  injury_model = c("bimodal", "separate", "none"),
  owner_efficiency = list(average = 0.75, sd = 0.025),
  verbose = FALSE
)
}

}
