% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_ff_wins_added.R
\name{ff_wins_added}
\alias{ff_wins_added}
\title{Wins Added}
\usage{
ff_wins_added(conn, ..., verbose = NULL)
}
\arguments{
\item{conn}{an connection to a league made with \code{ff_connect()} and friends (required)}

\item{...}{
  Arguments passed on to \code{\link[=ff_simulate]{ff_simulate}}
  \describe{
    \item{\code{n_seasons}}{number of seasons to simulate, default = 100}
    \item{\code{n_weeks}}{number of weeks per season, default = 14}
    \item{\code{best_ball}}{a logical: are weekly wins based on optimal lineups?}
    \item{\code{seed}}{an integer to control reproducibility}
    \item{\code{gp_model}}{select between "simple", "none" to apply a model for whether a player played in a given game, defaults to "simple"}
    \item{\code{base_seasons}}{a numeric vector that selects seasons as base data, earliest available is 2012}
    \item{\code{actual_schedule}}{a logical: use actual ff_schedule? default is FALSE}
    \item{\code{pos_filter}}{a character vector of positions to filter/run, default is c("QB","RB","WR","TE","K")}
    \item{\code{return}}{one of c("default", "all") - what objects to return in the output list}
  }}

\item{verbose}{a logical: print status messages? default is TRUE, configure with options(ffsimulator.verbose)}
}
\value{

}
\description{
Run base simulation once for n seasons/weeks etc
}
\details{
For every player on every team (except replacement level players),
\itemize{
\item remove them from that specific roster
\item reoptimize the lineups just for that roster without the player
\item use that set of weekly scores as the score for that week
\item summarise the new simulation
\item return new wins total and points total for that franchise
}

Summarise WA as the difference between the sim with the player and the sim without them
}
