% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_projections.R
\name{ffs_generate_projections}
\alias{ffs_generate_projections}
\title{Generate Projections}
\usage{
ffs_generate_projections(
  adp_outcomes,
  latest_rankings,
  n_seasons,
  n_weeks,
  rosters = NULL
)
}
\arguments{
\item{adp_outcomes}{a dataframe of adp-based weekly outcomes, as created by \code{ffs_adp_outcomes()}}

\item{latest_rankings}{a dataframe of rankings, as created by \code{ffs_latest_rankings()}}

\item{n_seasons}{number of seasons}

\item{n_weeks}{weeks per season}

\item{rosters}{a dataframe of rosters, as created by \code{ffs_rosters()} - optional, reduces computation to just rostered players}
}
\value{

}
\description{
Run the bootstrapped resampling of player week outcomes on the latest rankings and rosters for a given number of seasons and weeks per season.
}
\details{
Dataframe should contain ecr, prob_gp, week_outcomes, injury_model, franchise_id, pos, and bye
}
