% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_preprocessing.R
\name{ffs_preprocess_data}
\alias{ffs_preprocess_data}
\alias{ffs_preprocess_data.mfl_conn}
\alias{ffs_preprocess_data.sleeper_conn}
\alias{ffs_preprocess_data.espn_conn}
\alias{ffs_preprocess_data.flea_conn}
\title{Preprocess data for simulation}
\usage{
ffs_preprocess_data(conn, rosters, latest_rankings, adp_outcomes)

\method{ffs_preprocess_data}{mfl_conn}(conn, rosters, latest_rankings, adp_outcomes)

\method{ffs_preprocess_data}{sleeper_conn}(conn, rosters, latest_rankings, adp_outcomes)

\method{ffs_preprocess_data}{espn_conn}(conn, rosters, latest_rankings, adp_outcomes)

\method{ffs_preprocess_data}{flea_conn}(conn, rosters, latest_rankings, adp_outcomes)
}
\description{
Performs joins, cleaning, filtering.
}
