% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-rosters.R
\name{ffs_rosters}
\alias{ffs_rosters}
\alias{ffs_rosters.mfl_conn}
\alias{ffs_rosters.sleeper_conn}
\alias{ffs_rosters.flea_conn}
\alias{ffs_rosters.espn_conn}
\title{Get Rosters}
\usage{
ffs_rosters(conn)

\method{ffs_rosters}{mfl_conn}(conn)

\method{ffs_rosters}{sleeper_conn}(conn)

\method{ffs_rosters}{flea_conn}(conn)

\method{ffs_rosters}{espn_conn}(conn)
}
\arguments{
\item{conn}{a connection object as created by \code{ffscrapr::ff_connect()} and friends.}
}
\value{
a dataframe of rosters that includes a fantasypros_id column
}
\description{
This function lightly wraps \code{ffscrapr::ff_rosters()} and adds fantasypros_id, which is a required column for ffsimulator.
}
\examples{
\donttest{
# cached examples
conn <- .ffs_cache_example("mfl_conn.rds")

try({ # prevents CRAN connectivity issues, not actually required in normal usage
  ffs_rosters(conn)
})
}

}
\seealso{
vignette("custom") for more detailed example usage
}
