% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_roster_scores.R
\name{ffs_score_rosters}
\alias{ffs_score_rosters}
\title{Join Rosters to Projected Scores}
\usage{
ffs_score_rosters(projected_scores, rosters)
}
\arguments{
\item{projected_scores}{a dataframe of projected scores, as created by \code{ffs_generate_projections()}}

\item{rosters}{a dataframe of rosters, as created by \code{ffs_rosters()}}
}
\value{
A dataframe of roster-level projected scores
}
\description{
Attaches projected scores to rosters (via a left-join) and creates a positional ranking column.
}
\examples{
\donttest{
# cached examples
projected_scores <- .ffs_cache("projected_scores.rds")
rosters <- .ffs_cache("mfl_rosters.rds")

ffs_score_rosters(projected_scores, rosters)
}

}
\seealso{
vignette("Custom Simulations") for example usage
}
