% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{package}
\name{ffsimulator-package}
\alias{ffsimulator}
\alias{ffsimulator-package}
\title{ffsimulator: Simulate Fantasy Football Seasons}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Uses bootstrap resampling to run fantasy football season simulations supported by historical rankings and 'nflfastR' data, calculating optimal lineups, and returning aggregated results.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ffsimulator.ffverse.com}
  \item \url{https://github.com/ffverse/ffsimulator}
  \item Report bugs at \url{https://github.com/ffverse/ffsimulator/issues}
}

}
\author{
\strong{Maintainer}: Tan Ho \email{tan@tanho.ca} (\href{https://orcid.org/0000-0001-8388-5155}{ORCID}) [copyright holder]

}
\keyword{internal}
