% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_outcomes.R
\docType{data}
\name{fp_injury_table}
\alias{fp_injury_table}
\title{Injury table}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 645 rows and 3 columns.
}
\usage{
fp_injury_table
}
\description{
This dataframe gives the probability of a given position/rank playing in a given game.
}
\keyword{datasets}
