% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data.R
\docType{data}
\name{fp_injury_table}
\alias{fp_injury_table}
\title{FP injury table}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 692 rows and 3 columns.
}
\usage{
fp_injury_table
}
\description{
This dataframe contains a column (\code{prob_gp}) for each positional ranking that describes the probability of a player with that preseason ADP playing in a given game. It is modelled from historical rankings data and the number of games played per season for a given positional rank.
}
\keyword{datasets}
