% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data.R
\name{fp_rankings_history}
\alias{fp_rankings_history}
\title{Historical draft position ranks}
\usage{
fp_rankings_history()
}
\value{
dataframe of preseason/draft rankings from ffpros
}
\description{
This dataframe has historical positional draft rankings starting in 2012 for
QB/RB/WR/TE/PK and 2015 DL/LB/DB, as gathered by the ffpros package.
}
\details{
By default, it returns bundled package data unless the \code{ffsimulator.cache_directory}
option is configured, in which case it will look for a rds file named
\code{fp_rankings_history.rds} in that directory.
}
