% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data.R
\docType{data}
\name{fp_rankings_history_week}
\alias{fp_rankings_history_week}
\title{Historical position ranks}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 94257 rows and 11 columns.
}
\usage{
fp_rankings_history_week
}
\description{
This dataframe has historical positional in-season rankings for 2012-2020 QB/RB/WR/TE/PK and 2015-2020 DL/LB/DB, as gathered by the ffpros package.
}
\keyword{datasets}
